requirements.txt
boto3==1.24.70
botocore==1.27.70
jmespath==1.0.1
Pillow==6.0.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
